requirements.txt
  
asgiref==3.2.7            # via django
attrs==19.3.0             # via pytest
certifi==2020.4.5.1       # via requests
chardet==3.0.4            # via requests
django==3.0.5             # via -r requirements.in, djangorestframework
django-filter==2.4.0      # MyFilter
djangorestframework==3.11.0  # via -r requirements.in
djangorestframework-simplejwt==4.6.0 #
idna==2.9                 # via requests
importlib-metadata==1.6.0  # via pluggy, pytest
gunicorn==20.0.4
more-itertools==8.2.0     # via pytest
packaging==20.3           # via pytest
pluggy==0.13.1            # via pytest
psycopg2-binary==2.8.5
py==1.8.1                 # via pytest
PyJWT==2.0.1		  # 
pyparsing==2.4.7          # via packaging
pytest-django==3.9.0      # via -r requirements.in
pytest==5.4.1             # via pytest-django
pytz==2019.3              # via django
requests==2.23.0          # via -r requirements.in
six==1.14.0               # via packaging
sqlparse==0.3.1           # via django
urllib3==1.25.9           # via requests
wcwidth==0.1.9            # via pytest
zipp==3.1.0               # via importlib-metadata
python-dotenv==0.15.0